GRIT: Allow grd files outside of chromium source dir.
When replacing substrings with values from a dictionary,
the replacements should not be truncated to the length of src_root_dir.

If the basename does not exist as a key in the resource_ids file,
then fall back to an absolute path.

The precedence is like:
file in chromium > file basename > absolute file name.

THIS SHOULD BE UPSTREAMABLE.